+++ Adj/Program/Controlflow/Implementation/State.hs
@@ -7,3 +7,6 @@ type State state = (->) state =!?= (:*:) state
 
 current :: State state state
 current = FG .: \state -> state :*: state
+
+modify :: (state -> state) -> State state state
+modify m = FG .: \state -> m state :*: state