+++ Adj/Program/Controlflow/Implementation.hs
@@ -2,3 +2,4 @@ module Adj.Program.Controlflow.Implementation (module Exports) where
 
 import Adj.Program.Controlflow.Implementation.State as Exports
 import Adj.Program.Controlflow.Implementation.Store as Exports
+import Adj.Program.Controlflow.Implementation.Optics as Exports