+++ Adj/Program/Controlflow/Implementation/Optics.hs
@@ -1,0 +1,3 @@
+module Adj.Program.Controlflow.Implementation.Optics where
+
+-- We can recreate P_Q_T from Pandora with Flat and Dual wrappers