+++ Adj/Program/Controlflow/Implementation/Optics.hs
@@ -16,3 +16,7 @@ review :: Prism available set subset -> set -> available subset
 review prism set = case prism =- set of
 	FG (Flat (This subset)) -> subset
 	FG (Flat (That m)) -> m =- set
+
+-- TODO: think about composition between different type of optics
+-- TODO: Lens and Prism types looks the same so maybe we can just generalize them somehow
+-- TODO: think about cardinality in a target of Lens