+++ CHANGELOG.md
@@ -158,3 +158,7 @@
 # 0.0.8
 * Change `Generation` definition
 * Define `||` union constraint type from @rampion's package
+* Remove `.:+:` type alias
+* Define `+:` type alias
+* Remove `:+:.` type alias
+* Define `:+` type alias