+++ CHANGELOG.md
@@ -162,3 +162,7 @@
 * Define `+:` type alias
 * Remove `:+:.` type alias
 * Define `:+` type alias
+* Remove `.:*:` type alias
+* Define `*:` type alias
+* Remove `:*:.` type alias
+* Define `:*` type alias