+++ CHANGELOG.md
@@ -166,3 +166,5 @@
 * Define `*:` type alias
 * Remove `:*:.` type alias
 * Define `:*` type alias
+* Define `*|` operator
+* Define `|*` operator