+++ CHANGELOG.md
@@ -169,3 +169,4 @@
 * Rename `Option` module and type to `Maybe`
 * Change `Maybe` type definition
 * Change `Progress` type definition
+* Define `Adj.Program.Controlflow.Implementation.Store` module