+++ CHANGELOG.md
@@ -174,3 +174,4 @@
 * Define `modify` method
 * Define `position` method
 * Define `retrofit` method
+* Define `Adj.Program.Controlflow.Implementation.Optics` module