+++ CHANGELOG.md
@@ -176,3 +176,4 @@
 * Define `retrofit` method
 * Define `Adj.Program.Controlflow.Implementation.Optics` module
 * Change `State` type definition
+* Define `Lens` type