+++ CHANGELOG.md
@@ -177,3 +177,4 @@
 * Define `Adj.Program.Controlflow.Implementation.Optics` module
 * Change `State` type definition
 * Define `Lens` type
+* Define `view` method